add visualisation script to cli
