+++ .gitignore
@@ -5,3 +5,5 @@ lightning_logs
 checkpoints
 spares
 output
+*.png
+workbench
